Refactor build configs and improve type compatibility
Improved build configuration by adding `OutDir` properties to
multiple project files for consistent output directory management
across `Debug|x64` and `Release|x64` configurations.

Updated `Main.cpp` to use `ULONGLONG` for `last_stats_time` and
`current_time` to ensure compatibility with `GetTickCount64`.

Streamlined project dependencies by removing unnecessary
dependencies in `CryptoShield.sln` and adding a `ProjectReference`
to link `CryptoShieldService` with `CryptoShieldCore`.

These changes enhance build consistency, maintainability, and
runtime correctness.

+++ Service/CryptoShieldService/Main.cpp
@@ -379,7 +379,7 @@ DWORD WINAPI ServiceWorkerThread(LPVOID lpParam)
     WriteEventLog(EVENTLOG_INFORMATION_TYPE, L"Service worker thread started");
 
     // Main service loop
-    DWORD last_stats_time = GetTickCount64();
+    ULONGLONG last_stats_time = GetTickCount64();
     while (g_running) {
         if (!g_paused) {
             // Check connection status
@@ -397,7 +397,7 @@ DWORD WINAPI ServiceWorkerThread(LPVOID lpParam)
             }
 
             // Log statistics every minute
-            DWORD current_time = GetTickCount64();
+            ULONGLONG current_time = GetTickCount64();
             if (current_time - last_stats_time > 60000) {
                 auto stats = message_processor->GetStatistics();
                 std::wstring stats_msg = L"Statistics - Total operations: " +
